health-monitor/requirements.txt
google-cloud-servicehealth==0.1.3
google-cloud-firestore==2.14.0
python-dotenv==1.0.0
